+++ src/Service/BasketService.php
@@ -8,16 +8,28 @@ declare(strict_types = 1);
 namespace WizaplaceFrontBundle\Service;
 
 use GuzzleHttp\Exception\ClientException;
+use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
+use Psr\Log\NullLogger;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\AuthenticationEvents;
+use Symfony\Component\Security\Core\Event\AuthenticationEvent;
+use Symfony\Component\Security\Http\Logout\LogoutHandlerInterface;
+use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Basket\Basket;
 use Wizaplace\SDK\Basket\Comment;
 use Wizaplace\SDK\Basket\PaymentInformation;
 use Wizaplace\SDK\Catalog\DeclinationId;
+use WizaplaceFrontBundle\Security\User;
 
 /**
  * Wraps {@see \Wizaplace\SDK\Basket\BasketService}, storing the basketID for you.
  */
-class BasketService
+class BasketService implements EventSubscriberInterface, LogoutHandlerInterface
 {
     private const ID_SESSION_KEY = '_basketId';
 
@@ -30,16 +42,28 @@ class BasketService
     /** @var null|Basket */
     private $basket;
 
-    public function __construct(\Wizaplace\SDK\Basket\BasketService $baseService, SessionInterface $session)
+    /** @var LoggerInterface */
+    private $logger;
+
+    public function __construct(\Wizaplace\SDK\Basket\BasketService $baseService, SessionInterface $session, ?LoggerInterface $logger = null)
     {
         $this->baseService = $baseService;
         $this->session = $session;
+
+        if ($logger === null) {
+            $logger = new NullLogger();
+        }
+        $this->logger = $logger;
     }
 
     public function getBasket(): Basket
     {
-        $basketId = $this->getBasketId();
-        if (!$this->basket) {
+        $basketId = $this->getCurrentBasketId();
+        if ($basketId === null) {
+            return Basket::createEmpty('fake-empty-basket');
+        }
+
+        if (!$this->basket || $this->basket->getId() !== $basketId) {
             try {
                 $this->basket = $this->baseService->getBasket($basketId);
             } catch (ClientException $e) {
@@ -54,49 +78,90 @@ class BasketService
         return $this->basket;
     }
 
+    /**
+     * @throws \Wizaplace\SDK\Basket\Exception\BadQuantity
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     */
     public function addProductToBasket(DeclinationId $declinationId, int $quantity): int
     {
-        $this->basket = null;
+        try {
+            $result = $this->baseService->addProductToBasket($this->getBasketId(), $declinationId, $quantity);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
 
-        return $this->baseService->addProductToBasket($this->getBasketId(), $declinationId, $quantity);
+        return $result;
     }
 
+    /**
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     */
     public function removeProductFromBasket(DeclinationId $declinationId): void
     {
-        $this->basket = null;
-
-        $this->baseService->removeProductFromBasket($this->getBasketId(), $declinationId);
+        try {
+            $this->baseService->removeProductFromBasket($this->getBasketId(), $declinationId);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
     }
 
+    /**
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     */
     public function cleanBasket(): void
     {
-        $this->basket = null;
-
-        $this->baseService->cleanBasket($this->getBasketId());
+        try {
+            $this->baseService->cleanBasket($this->getBasketId());
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
     }
 
+    /**
+     * @throws \Wizaplace\SDK\Basket\Exception\BadQuantity
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     */
     public function updateProductQuantity(DeclinationId $declinationId, int $quantity): int
     {
-        $this->basket = null;
+        try {
+            $result = $this->baseService->updateProductQuantity($this->getBasketId(), $declinationId, $quantity);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
 
-        return $this->baseService->updateProductQuantity($this->getBasketId(), $declinationId, $quantity);
+        return $result;
     }
 
+    /**
+     * @throws \Wizaplace\SDK\Basket\Exception\CouponAlreadyPresent
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     */
     public function addCoupon(string $coupon): void
     {
-        $this->basket = null;
-        $this->baseService->addCoupon($this->getBasketId(), $coupon);
+        try {
+            $this->baseService->addCoupon($this->getBasketId(), $coupon);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
     }
 
+    /**
+     * @throws \Wizaplace\SDK\Basket\Exception\CouponNotInTheBasket
+     */
     public function removeCoupon(string $coupon): void
     {
-        $this->basket = null;
-        $this->baseService->removeCoupon($this->getBasketId(), $coupon);
+        try {
+            $this->baseService->removeCoupon($this->getBasketId(), $coupon);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
     }
 
     /**
      * @see \Wizaplace\SDK\Basket\BasketService::getPayments
      * @return \Wizaplace\SDK\Basket\Payment[]
+     * @throws AuthenticationRequired
+     * @throws \Wizaplace\SDK\Exception\NotFound
      */
     public function getPayments(): array
     {
@@ -105,29 +170,100 @@ class BasketService
 
     public function selectShippings(array $selections): void
     {
-        $this->basket = null;
-        $this->baseService->selectShippings($this->getBasketId(), $selections);
+        try {
+            $this->baseService->selectShippings($this->getBasketId(), $selections);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
     }
 
+    /**
+     * @throws AuthenticationRequired
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     * @throws \Wizaplace\SDK\Exception\SomeParametersAreInvalid
+     */
     public function checkout(int $paymentId, bool $acceptTerms, string $redirectUrl): PaymentInformation
     {
-        $this->basket = null;
+        try {
+            $result = $this->baseService->checkout($this->getBasketId(), $paymentId, $acceptTerms, $redirectUrl);
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
 
-        return $this->baseService->checkout($this->getBasketId(), $paymentId, $acceptTerms, $redirectUrl);
+        return $result;
     }
 
     public function forgetBasket(): void
     {
-        $this->basket = null;
-        $this->session->remove(self::ID_SESSION_KEY);
+        try {
+            $this->session->remove(self::ID_SESSION_KEY);
+            try {
+                $this->baseService->setUserBasketId(null);
+            } catch (AuthenticationRequired $e) {
+                // We are not logged in, this is not a real error.
+            }
+        } finally {
+            $this->basket = null; // invalidate local cache
+        }
     }
 
     /**
      * @param $comments Comment[]
+     * @throws \Wizaplace\SDK\Exception\NotFound
+     * @throws \Wizaplace\SDK\Exception\SomeParametersAreInvalid
      */
     public function updateComments(array $comments): void
     {
         $this->baseService->updateComments($this->getBasketId(), $comments);
+    }
+
+    public function onAuthenticationSuccess(AuthenticationEvent $event): void
+    {
+        $user = $event->getAuthenticationToken()->getUser();
+        if (!($user instanceof User)) {
+            return;
+        }
+
+        try {
+            $userBasketId = $this->baseService->getUserBasketId();
+
+            if ($userBasketId === null) {
+                $this->baseService->setUserBasketId($this->getBasketId());
+
+                return;
+            }
+
+            $currentBasketId = $this->getCurrentBasketId();
+            if (null !== $currentBasketId) {
+                $this->baseService->mergeBaskets($userBasketId, $currentBasketId);
+            }
+
+            $this->setCurrentBasketId($userBasketId);
+        } catch (\Throwable $e) {
+            // We just log the exception, we don't want this to cause an error page.
+            $this->logger->log(LogLevel::ERROR, 'Failed to load or merge the user\'s basket ID', [
+                'exception' => $e,
+            ]);
+        }
+    }
+
+    /**
+     * @inheritdoc
+     * @uses \WizaplaceFrontBundle\Service\BasketService::onAuthenticationSuccess
+     */
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            AuthenticationEvents::AUTHENTICATION_SUCCESS => 'onAuthenticationSuccess',
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function logout(Request $request, Response $response, TokenInterface $token): void
+    {
+        $this->forgetBasket();
     }
 
     /**
@@ -136,13 +272,30 @@ class BasketService
      */
     private function getBasketId(): string
     {
-        $basketId = $this->session->get(self::ID_SESSION_KEY);
+        $basketId = $this->getCurrentBasketId();
 
         if (null === $basketId) {
-            $basketId = $this->baseService->create();
-            $this->session->set(self::ID_SESSION_KEY, $basketId);
+            $this->basket = $this->baseService->createEmptyBasket();
+            $basketId = $this->basket->getId();
+            $this->setCurrentBasketId($basketId);
         }
 
         return $basketId;
+    }
+
+
+    /**
+     * Gets current basket ID, if it exists.
+     * Most of the time you should not use this , {@see \WizaplaceFrontBundle\Service\BasketService::getBasketId} instead.
+     * @return null|string
+     */
+    private function getCurrentBasketId(): ?string
+    {
+        return $this->session->get(self::ID_SESSION_KEY, null);
+    }
+
+    private function setCurrentBasketId(string $basketId): void
+    {
+        $this->session->set(self::ID_SESSION_KEY, $basketId);
     }
 }